+++ packages/shadow-objects/src/view/cloneChangeTrail.ts
@@ -4,7 +4,7 @@ export function cloneChangeTrail(trails: ChangeTrailType): ChangeTrailType {
   return trails.map((trail) => {
     if (trail.transferables && trail.transferables.length > 0) {
       const {transferables, ...data} = trail;
-      return structuredClone(data, {transfer: transferables as any});
+      return structuredClone(data, {transfer: transferables});
     } else {
       return structuredClone(trail);
     }